# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Các API liên quan đến User.",
        "title": "Service Auth",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/auth/create-user": {
            "post": {
                "description": "Tạo User",
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "description": "Thông tin tạo User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
        "/auth/get-user/{id}": {
            "get": {
                "description": "Lấy thông tin User theo ID",
                "tags": [
                    "admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
        "/auth/get-users": {
            "get": {
                "description": "Lấy danh sách User",
                "tags": [
                    "admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        },
        "/auth/update-user/{id}": {
            "put": {
                "description": "Cập nhật User theo ID",
                "tags": [
                    "admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "Ảnh đại diện",
                    "type": "string"
                },
                "email": {
                    "description": "Địa chỉ email",
                    "type": "string"
                },
                "full_name": {
                    "description": "Tên hiển thị",
                    "type": "string"
                },
                "password": {
                    "description": "Mật khẩu phải có tối thiểu 8 ký tự",
                    "type": "string"
                },
                "phone": {
                    "description": "Số điện thoại",
                    "type": "string"
                },
                "roles": {
                    "description": "Mảng các quyền hạn (Role) của User (ví dụ [3, 5] tức là User có Role Id là 3 và 5).\nMột User có thể có nhiều Role, nếu Null tức là User không có quyền gì",
                    "type": "array",
                    "items": {
                        "type": "int32"
                    }
                }
            }
        },
        "model.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}